Revert "Change Time to uint64"
This reverts commit 57f45a562f11ceebe2eb63038052e532f1bfff27.

+++ hcx/network_profile.go
@@ -55,7 +55,7 @@ type NetworkIpRange struct {
 type NetworkProfileResult struct {
 	Success   bool               `json:"success"`
 	Completed bool               `json:"completed"`
-	Time      uint64             `json:"time"`
+	Time      int                `json:"time"`
 	Data      NetworkProfileData `json:"data"`
 }
